# logisland-framework/logisland-agent/src/main/swagger/api-swagger.yaml
swagger: '2.0'
info:
  description: REST API for logisland agent
  version: v1
  title: logisland-agent
  contact:
    name: Thomas Bailet
    email: bailet.thomas@gmail.com
host: localhost:8081
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /jobs:
    get:
      tags:
       - job
      operationId: getAllJobs
      summary:  get all jobs
      description: retrieve all jobs (retrieve only summary fields)
      responses:
        "200":
          description: job configuration list
          schema:
            type: array
            items:
              $ref: '#/definitions/Job'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      tags:
       - job
      summary:  create new job
      description:  store a new job configuration if valid
      operationId: addJob
      parameters:
        - name: job
          in: body
          description: Job to add to the store
          required: true
          schema:
            $ref: '#/definitions/Job'
      responses:
        "404":
          description: Job not found
        "400":
          description: Invalid ID supplied
        "200":
          description: Job successfuly created
          schema:
            $ref: '#/definitions/Job'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /jobs/alerts:
     get:
      tags:
        - job
      summary: get job alerts
      description: get the alerts
      operationId: getJobAlerts
      parameters:
        - name: count
          in: query
          description: max number of ites to retrieve
          required: false
          type: integer
          default: 20
      responses:
        "200":
          description: job metrics
          schema:
            type: array
            items:
              $ref: '#/definitions/Record'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /jobs/errors:
     get:
      tags:
        - job
      summary: get last job errors
      description: get the errors
      operationId: getJobErrors
      parameters:
        - name: count
          in: query
          description: max number of ites to retrieve
          required: false
          type: integer
          default: 20
      responses:
        "200":
          description: job errors
          schema:
            $ref: '#/definitions/Job'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /jobs/{jobId}:
    get:
      tags:
        - job
      summary: get job
      description: get the corresponding Job definition
      operationId: getJob
      produces:
        - application/json
        - text/plain
      parameters:
        - name: jobId
          in: path
          description: id of the job to return
          required: true
          type: string
      responses:
        "200":
          description: job definition
          schema:
            $ref: '#/definitions/Job'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - job
      summary: delete job
      description: remove the corresponding Job definition and stop if its currently running
      operationId: deleteJob
      parameters:
        - name: jobId
          in: path
          description: id of the job to return
          required: true
          type: string
      responses:
        "404":
          description: Job not found
        "400":
          description: Invalid ID supplied
        "200":
          description: job successfully removed
          schema:
            $ref: '#/definitions/Job'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
       - job
      summary:  update job
      description: update an existing job configuration if valid
      operationId: updateJob
      parameters:
        - name: jobId
          in: path
          description: Job to add to the store
          required: true
          type: string
        - name: job
          in: body
          description: Job to add to the store
          required: true
          schema:
            $ref: '#/definitions/Job'
      responses:
        "200":
          description: Job successfuly created
          schema:
            $ref: '#/definitions/Job'
    post:
      tags:
       - job
      summary:  create new job
      description:  store a new job configuration if valid
      operationId: addJobWithId
      parameters:
        - name: body
          in: body
          description: Job configuration to add to the store
          required: true
          schema:
            $ref: '#/definitions/Job'
        - name: jobId
          in: path
          description: JobId to add to the store
          required: true
          type: string
      responses:
        "404":
          description: Job not found
        "400":
          description: Invalid ID supplied
        "200":
          description: Job successfuly created
          schema:
            $ref: '#/definitions/Job'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /jobs/{jobId}/start:
     post:
      tags:
        - job
      summary: start job
      description: start the corresponding Job definition
      operationId: startJob
      parameters:
        - name: jobId
          in: path
          description: id of the job to return
          required: true
          type: string
      responses:
        "200":
          description: job successfuly started
          schema:
            $ref: '#/definitions/Job'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /jobs/{jobId}/shutdown:
     post:
      tags:
        - job
      summary: shutdown job
      description: shutdown the running Job
      operationId: shutdownJob
      parameters:
        - name: jobId
          in: path
          description: id of the job to return
          required: true
          type: string
      responses:
        "200":
          description: job successfuly started
          schema:
            $ref: '#/definitions/Job'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /jobs/{jobId}/pause:
     post:
      tags:
        - job
      summary: pause job
      description: pause the corresponding Job
      operationId: pauseJob
      parameters:
        - name: jobId
          in: path
          description: id of the job to return
          required: true
          type: string
      responses:
        "200":
          description: job successfuly paused
          schema:
            $ref: '#/definitions/Job'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /jobs/{jobId}/status:
     get:
      tags:
        - job
      summary: get job status
      description: get the status of corresponding Job
      operationId: getJobStatus
      parameters:
        - name: jobId
          in: path
          description: id of the job to return
          required: true
          type: string
      responses:
        "200":
          description: job status
          schema:
            $ref: '#/definitions/Job'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'


  # Topics API
...
definitions:
  Field:
    type: object
    required:
      - name
      - type
    properties:
      name:
        description: a unique identifier for the topic
        type: string
      encrypted:
        description: is the field need to be encrypted
        type: boolean
      indexed:
        description: is the field need to be indexed to search store
        type: boolean
      persistent:
        description: is the field need to be persisted to data store
        type: boolean
      type:
        description: the type of the field
        type: string
        default: string
        enum:
        - null
        - string
        - long
        - array
        - float
        - double
        - bytes
        - record
        - map
        - enum
        - boolean
      value:
        description: a unique identifier for the topic
        type: object
  Record:
    type: object
    required:
      - name
    properties:
      id:
        description: a unique identifier for the topic
        type: integer
        format: int64
      name:
        description: the name of the topic
        type: string
      multiline:
        description: is the record multiline
        type: boolean
      timestampField:
        description: the record_time field
        type: string
        default: "record_time"
      rowkeyField:
        description: the record_id field
        type: string
        default: "record_id"
      fields:
        type: array
        items:
          $ref: '#/definitions/Field'
...
  JobSummary:
    type: object
    properties:
      usedCores:
        description: the number of used cores
        type: integer
        format: int32
      usedMemory:
        description: the total memory allocated for this job
        type: integer
        format: int32
      status:
        description: the job status
        type: string
        default: stopped
        enum:
        - stopped
        - running
        - failed
      dateModified:
        description: latest date of modification
        type: string
        format: date-time
      documentation:
        description: write here what the job is doing
        type: string
  Job:
    type: object
    required:
      - engine
      - name
      - streams
      - version
    properties:
      id:
        description: a unique identifier for the job
        type: integer
        format: int64
      version:
        description: the version of the job configuration
        type: integer
        format: int32
      name:
        description: the job name
        type: string
      engine:
        $ref: '#/definitions/Engine'
      streams:
        type: array
        items:
          $ref: '#/definitions/Stream'
      summary:
        $ref: '#/definitions/JobSummary'
  Engine:
    type: object
    required:
      - component
      - config
      - name
    properties:
      name:
        type: string
      component:
        type: string
      documentation:
        type: string
      config:
        type: array
        items:
          $ref: '#/definitions/Property'
  Stream:
    type: object
    required:
      - component
      - name
    properties:
      name:
        type: string
      component:
        type: string
      documentation:
        type: string
      config:
        type: array
        items:
          $ref: '#/definitions/Property'
      processors:
        type: array
        items:
          $ref: '#/definitions/Processor'
  Processor:
    type: object
    required:
      - component
      - config
      - name
    properties:
      name:
        type: string
      component:
        type: string
      documentation:
        type: string
      config:
        type: array
        items:
          $ref: '#/definitions/Property'
  Property:
    type: object
    required:
      - key
      - value
    properties:
      key:
        type: string
      type:
        type: string
      value:
        type: string
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string